Split interops into specific files
